+++ 2023/day03/src/part1.rs
@@ -3,7 +3,7 @@ use crate::{aoc_error::AocError, parser::parse, Engine, EngineMap};
 pub fn process(input: &str) -> miette::Result<String, AocError> {
     let engine: Engine = parse(input)?;
     let map: EngineMap = engine.into();
-    let parts = map.get_parts_next_to_symbols();
+    let parts = map.get_parts_neighbouring_any_symbol();
     let total: usize = parts.iter().map(|part| part.id).sum();
     Ok(total.to_string())
 }
